Refactor main.py #949 - code split in few files, all Python code moved to app/ directory, helper functions in separate file, routes in separate files, init file for creating and configuring the Flask app
